requirements.txt
discord.py>=2.0.0
firebase-admin>=6.0.0 